migrate ConsumerReportReadByProNotification
