+++ app/services/emails/EmailDefinitionsConsumer.scala
@@ -247,6 +247,35 @@ object EmailDefinitionsConsumer {
     }
   }
 
+  case object ConsumerReportReadByProNotification extends EmailDefinition {
+    override val category = Consumer
+
+    override def examples =
+      Seq(
+        "report_transmitted" -> ((recipient, messagesApi) =>
+          EmailImpl(genReport.copy(email = recipient), Some(genCompany), messagesApi)
+        )
+      )
+
+    final case class EmailImpl(
+        report: Report,
+        maybeCompany: Option[Company],
+        messagesApi: MessagesApi
+    ) extends ConsumerEmail {
+      private val lang                                        = Lang(getLocaleOrDefault(report.lang))
+      implicit private val messagesProvider: MessagesProvider = MessagesImpl(lang, messagesApi)
+
+      override val recipients: List[EmailAddress] = List(report.email)
+      override val subject: String                = messagesApi("ConsumerReportTransmittedEmail.subject")(lang)
+
+      override def getBody: (FrontRoute, EmailAddress) => String = (_, _) =>
+        views.html.mails.consumer.reportTransmission(report, maybeCompany).toString
+
+      override def getAttachements: AttachmentService => Seq[Attachment] =
+        _.attachmentSeqForWorkflowStepN(3, report.lang.getOrElse(Locale.FRENCH))
+    }
+  }
+
   private def getLocaleOrDefault(locale: Option[Locale]): Locale = locale.getOrElse(Locale.FRENCH)
 
 }
